Partial DataboxController Refactor
